Move deck limit
Changing deck limit for move cards to 3.  This is because you might want to build a deck with 2 copies of that card plus 1 copy under lightsaber mastery (you can't have two copies under lightsaber mastery).  You might have a more elegant solution in mind, but this is a dix that at least allows you to build the deck you want for now.