Update to SoH.json
Cards spoiled on 6/28 (EST).  I believe everything's been formatted correctly for the json.  I know I've missed some of the reminder text in the "text" key for the ([melee])([ranged])([indirect]).  But I believe a majority of the data is in.   However, it's 2:30am EST and I need sleep.  I'll pick up the corrections this weekend I.